requirements.txt
amqp==2.5.0
anyjson==0.3.3
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
billiard==3.5.0.5
botocore==1.12.229
cachetools==3.1.1
celery==4.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.3.9
coverage==4.5.4
dateparser==0.7.2
decorator==4.4.0
dj-database-url==0.5.0
Django==1.11.23
django-cors-headers==3.2.0
django-prometheus-metrics==0.0.1
django-tables2==1.16.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
docutils==0.15.2
EasyProcess==0.2.7
entrypoints==0.3
flake8==3.7.8
future==0.17.1
google-auth==1.7.1
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
ipdb==0.12
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.0
jmespath==0.9.4
kombu==4.1.0
kubernetes==10.0.1
mccabe==0.6.1
more-itertools==7.2.0
oauthlib==3.0.2
packaging==19.2
parso==0.5.0
pendulum==2.0.5
pexpect==4.7.0
pickleshare==0.7.5
pika==1.1.0
Pillow==6.2.0
pluggy==0.13.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
python-dateutil==2.8.0
python-json-logger==0.1.11
pytz==2019.1
pytzdata==2019.3
PyYAML==5.1.1
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
traitlets==4.3.2
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
websocket-client==0.56.0
yapf==0.27.0
zipp==0.6.0
